Fixed notest loops not iterating all the way

+++ prng_runner.c
@@ -80,7 +80,7 @@ void experiment_prng(char* prng)
                nrandnums);
         fflush(stdout);
 
-        for(int i = 0; i < nrandnums; i++)
+        for(unsigned long i = 0; i < nrandnums; i++)
         {
             xorshift32(&xorshift32_seed, &shift1, &shift2, &shift3);
         }
@@ -90,7 +90,7 @@ void experiment_prng(char* prng)
         printf("Generating %llu random numbers with lfg \n", nrandnums);
         fflush(stdout);
 
-        for(int i = 0; i < nrandnums; i++)
+        for(unsigned long i = 0; i < nrandnums; i++)
         {
             lfg(&lag1, &lag2, &modval, initvals);
         }
@@ -100,7 +100,7 @@ void experiment_prng(char* prng)
         printf("Generating %llu random numbers with lcg \n", nrandnums);
         fflush(stdout);
 
-        for(int i = 0; i < nrandnums; i++)
+        for(unsigned long i = 0; i < nrandnums; i++)
         {
             lcg(&lcg_seed, &modval, &multfactor, &addfactor);
         }
